requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
openai
langchain
langgraph
streamlit
httpx
python-multipart
typing-extensions